Rename package to @next/env
