+++ packages/next-env/README.md
@@ -1,3 +1,3 @@
-# `@next/dotenv-loading`
+# `@next/env`
 
 Next.js' util for loading dotenv files in with the proper priorities
