+++ packages/next-env/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@next/dotenv-loading",
+  "name": "@next/env",
   "version": "9.5.4-canary.20",
   "keywords": [
     "react",
@@ -11,7 +11,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/vercel/next.js",
-    "directory": "packages/next-dotenv-loading"
+    "directory": "packages/next-env"
   },
   "author": "Next.js Team <support@vercel.com>",
   "license": "MIT",
